make detach interruptible and restartable (#279)
Run detach as a work item from system queue since it can generate io and
block. Wait the detach completion interruptible so the process can be
killed while in process of detach.

Keep the device on the context list while detach work is in progress so
new callers can find it and block until completion.

Signed-off-by: Lakshmi Narasimhan Sundararajan <lsundararajan@purestorage.com>